move oauth provider to common folder
