+++ busy_beaver/app.py
@@ -1,6 +1,6 @@
 from flask import Flask, request
 
-from .apps.oauth_integrations.oauth_providers.base import OAuthError
+from .common.oauth import OAuthError
 from .config import DATABASE_URI, REDIS_URI, SECRET_KEY
 from .extensions import db, migrate, rq, talisman
 from .exceptions import NotAuthorized, ValidationError
