+++ busy_beaver/apps/github_integration/workflow.py
@@ -1,7 +1,7 @@
 import logging
 from typing import NamedTuple
 
-from busy_beaver.apps.oauth_integrations.oauth_providers.base import OAuthError
+from busy_beaver.common.oauth import OAuthError
 from busy_beaver.clients import github_oauth
 from busy_beaver.extensions import db
 from busy_beaver.models import GitHubSummaryUser, SlackInstallation
