+++ busy_beaver/apps/oauth_integrations/oauth_providers/__init__.py
@@ -1,2 +1,1 @@
-from .base import ExternalOAuthDetails, OAuthError  # noqa
 from .slack import StateToOAuthResponse  # noqa
